fix: do not delete if generated filename is empty (#4528)

+++ packages/amplify-frontend-ios/lib/frontend-config-creator.js
@@ -35,8 +35,9 @@ function deleteAmplifyConfig(context) {
         const file = path.join(srcDirPath, docsFilePath, `${filename}.${FILE_EXTENSION_MAP[codeGenTarget]}`);
         if (fs.existsSync(file)) fs.removeSync(file);
       });
-
-      fs.removeSync(path.join(srcDirPath, generatedFileName));
+      if (generatedFileName.trim() !== '') {
+        fs.removeSync(path.join(srcDirPath, generatedFileName));
+      }
     });
   }
 }
